planet: test 16 shard build
I have a hypothesis that its important the total number of shards on a build and the
number of CPU cores on an Elasticsearch cluster are in alignment.

In particular the number of cores should be an _exact_ multiple of the
primary shard count.

CPU counts on AWS (and other) machines follow patterns like 4, 8, and 16
that are more aligned with a 16 shard build at most scales.

This change will test if that hypothesis is true.